Change plant data generation using Henneman data
This commit adds a utility function: load_plant_data. This function
replicates Henneman, et.al. (2018) code to obtain the total number of
polluting units in the US from 1995 to 2015. The data is uploaded to a
PostgreSQL database, but an option to store locally is also available.
Ideally this data will be used as a parameter source for modeling.