RPostgres Driver nightmare
This commit has a hotfix to deal with the schema name in the
dbWriteTable function. Long story short, there's no unique way to add
the schema name into the write table function. If we change drivers, we
will have to change the relationship of table and schema. The
documentation suggest the DBI::SQL function to unquote, but that has
proven ineffective.

We add a option into the engine driver to take the search path as the
schema.