Allow hyspslit to create pool connection [hotfix]
We were having troubles with the db connection pool inside hysplit. The
error was either signaling a segmentation error in the C libraries
underlying the pool library. Solve that is messy, so this commit calls
the mydb function to create the pool before uploading. The last approach
was more programatic and this task was done by the db_uploader function.

Let's see if Ubuntu likes this