Revamp dataabase strategy
Well, after playing around with DBI several times, this commit finally
introduces a more robust way of definying a schema into the DbWriteTable
function. DBI::Id creates a pointer that SQL can read in all the loops
and avoids several problems (i.e. not appending data).

Also, a more simplified way of uploading data is made. No more piping,
but creating an object in memory and sending it to the database using
DBI, is less nice and less tidyverse-like, but it makes the work in a
more robust and DBI-compliant way.

Parallel strategy is also changed by allowing a model directory names
with actual modeling parameters instead of the PID and current model
date. It is better than using the 6 miliseconds.